# packages/server/src/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.0",
    "title": "Boardie",
    "description": "API to serve the Boardie application"
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/me": {
      "post": {
        "summary": "Creates a new user",
        "tags": [
          "auth"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "middleAndSurname": {
                    "type": "string"
                  },
                  "age": {
                    "type": "number",
                    "minimum": 1,
                    "maximum": 120,
                    "nullable": true
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  },
                  "anonFriendInviteToken": {
                    "type": "string"
                  }
                },
                "required": [
                  "firstName",
                  "middleAndSurname",
                  "age",
                  "email",
                  "password"
                ],
                "additionalProperties": false,
                "description": "Data used to create a new user"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "firstName": {
                      "type": "string"
                    },
                    "middleAndSurname": {
                      "type": "string"
                    },
                    "age": {
                      "type": "number",
                      "minimum": 1,
                      "maximum": 120,
                      "nullable": true
                    },
                    "email": {
                      "type": "string",
                      "format": "email"
                    }
                  },
                  "required": [
                    "id",
                    "firstName",
                    "middleAndSurname",
                    "age",
                    "email"
                  ],
                  "additionalProperties": false,
                  "description": "Data that represents a user"
                }
              }
            }
          },
          "400": {
            "description": "The user data is invalid"
          },
          "403": {
            "description": "The anon-friend invite-token is invalid"
          },
          "409": {
            "description": "There's already a user registered with this email"
          }
        }
      },
      "get": {
        "summary": "Gets the logged user",
        "tags": [
          "auth"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "The logged user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "firstName": {
                      "type": "string"
                    },
                    "middleAndSurname": {
                      "type": "string"
                    },
                    "age": {
                      "type": "number",
                      "minimum": 1,
                      "maximum": 120,
                      "nullable": true
                    },
                    "email": {
                      "type": "string",
                      "format": "email"
                    }
                  },
                  "required": [
                    "id",
                    "firstName",
                    "middleAndSurname",
                    "age",
                    "email"
                  ],
                  "additionalProperties": false,
                  "description": "Data that represents a user"
                }
              }
            }
          },
          "404": {
            "description": "The logged user was not found"
          }
        }
      }
    },
    "/me/unregister": {
      "post": {
        "summary": "Unregisters the logged user",
        "tags": [
          "auth"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ],
                "additionalProperties": false,
                "description": "Auth information about a user"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The logged user was unregistered"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
...
  }
}